extract Advice and Asset for reuse in new tag cmd
