+++ helpers/asset/advice.go
@@ -1,0 +1,44 @@
+package asset
+
+import (
+	"fmt"
+
+	"github.com/simulot/immich-go/browser"
+	"github.com/simulot/immich-go/immich"
+)
+
+// - - go:generate stringer -type=AdviceCode
+type AdviceCode int
+
+func (a AdviceCode) String() string {
+	switch a {
+	case IDontKnow:
+		return "IDontKnow"
+	// case SameNameOnServerButNotSure:
+	// 	return "SameNameOnServerButNotSure"
+	case SmallerOnServer:
+		return "SmallerOnServer"
+	case BetterOnServer:
+		return "BetterOnServer"
+	case SameOnServer:
+		return "SameOnServer"
+	case NotOnServer:
+		return "NotOnServer"
+	}
+	return fmt.Sprintf("advice(%d)", a)
+}
+
+const (
+	IDontKnow AdviceCode = iota
+	SmallerOnServer
+	BetterOnServer
+	SameOnServer
+	NotOnServer
+)
+
+type Advice struct {
+	Advice      AdviceCode
+	Message     string
+	ServerAsset *immich.Asset
+	LocalAsset  *browser.LocalAssetFile
+}
